Trying new technique

+++ Week2/fibonacci_sum_last_digit_fast.cpp
@@ -11,7 +11,7 @@ int fibonacci_last_digit_fast(int n) {
 	uint64_t sum = 1;
 	int n2 = (n + 1 - 2) % 3;
 	int n1 = n + 1 - 2 - n2;
-	for (int i = 0; i < n1; i += 3) {
+	for (int64_t i = 0; i < n1; i += 3) {
 		f2 = f1 + f0;
 		sum = sum + f2;
 		f0 = f1 + f2;
@@ -31,7 +31,7 @@ int fibonacci_last_digit_fast(int n) {
 		sum = sum + f2;
 		f0 = f1; f1 = f2;
 	}
-	return sum%10;
+	return sum % 10;
 }
 
 int main() {
